Split marks into separate files
